style: 회원탈퇴화면 UI 반응형 개선
