+++ src/components/MobileWirtePostButton.tsx
@@ -1,0 +1,31 @@
+'use client';
+import { usePathname, useRouter } from 'next/navigation';
+
+export default function MobileWritePostButton() {
+    const router = useRouter();
+    const pathname = usePathname();
+
+    const moveToWritePost = () => {
+        const lastPathSegment = pathname.split('/').pop();
+
+        if (location.pathname.startsWith('/community/')) {
+            if (lastPathSegment && decodeURIComponent(lastPathSegment) !== '전체') {
+                router.push(`/writepost/${lastPathSegment}`);
+            } else {
+                router.push(`/writepost`);
+            }
+        } else {
+            router.push(`/writepost`);
+        }
+    };
+
+    return (
+        <button
+            type='button'
+            onClick={moveToWritePost}
+            className='aspect-square w-[5.2rem] bg-moneed-brand absolute bottom-[calc(100%+2rem)] flex items-center justify-center rounded-full right-8'
+        >
+            <img src='/icon/icon-edit.svg' alt='게시글 작성' />
+        </button>
+    );
+}
